+++ tests/FlightTest.php
@@ -6,9 +6,11 @@ if (! class_exists('PHPUnit_Framework_TestCase')) {
 
 class FlightTest extends PHPUnit_Framework_TestCase {
     private $CI;
+    private $plane;
+    
     public function setUp() {
         // Load CI instance normally
-        $this->CI = &get_instance();
+        $this -> CI = &get_instance();
     }
     public function testTrivial() {
         $this->assertEquals('bar', 'bar');