+++ changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2017-11-10 - SC
+### Added
+- Unit tests for Plane entity
+- function in WackyAPI to get albatros airline
+
 ## [Unreleased] - 2017-11-10 - YR
 ### Changed
 - Changed flightentity to contain airport objects